Split palette interaction test fixtures
