+++ Packages/macOS/CmuxCommandPalette/Tests/CmuxCommandPaletteTests/RecordingCommandPaletteObserverToken.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+final class RecordingCommandPaletteObserverToken: NSObject {
+    let id: Int
+
+    init(id: Int) {
+        self.id = id
+        super.init()
+    }
+}
