+++ Packages/macOS/CmuxCommandPalette/Tests/CmuxCommandPaletteTests/RecordingMouseDownView.swift
@@ -1,0 +1,10 @@
+import AppKit
+
+@MainActor
+final class RecordingMouseDownView: NSView {
+    private(set) var mouseDownCount = 0
+
+    override func mouseDown(with event: NSEvent) {
+        mouseDownCount += 1
+    }
+}
